Fixes
1) Do not add an extension if it's already there;
2) I've no idea why I've been using std::copy with std::back_inserter.
Changed that to operator=;
3) IV was a C-style array before the initial commit on GitHub, now it's
a vector too, thus I can simply use operator=;
4) Changed capacity() to not throw an exception, because returning a
zero is enough;
5) Defined some macros to make Windows.h less annoying;
6) Made custom implementation of zeroMemory inline.

+++ include/byteencryption.h
@@ -55,7 +55,7 @@ std::vector<uint8_t> SerpentDecrypt(const std::vector<uint8_t> &source, const by
  ** Uses 2nd template argument to determine iterations of Whirlpool
  ** (500 000 by default)
  **/
-template <int hashSize, int iterations = 500000>
+template <size_t hashSize, int iterations = 500000>
 std::array<byte, hashSize> hashKey(const std::string &key, const std::vector<byte> &salt) {
 	std::array<byte, hashSize> derived;
